+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # CHANGELOG
 https://keepachangelog.com/en/1.0.0/
 
+## [1.17.2] - 2023-01-02
+- `SynchronousTaskWait`: Now works for top-level statements
+- `SynchronousTaskWait`: Rewritten to use `IOperation`
+- `SynchronousTaskWait`: Preserves leading trivia when applying the code fix
+- `SynchronousTaskWait`: Doesn't offer a code fix when a timeout is passed to `Task.Wait(timeoutMs)`
+- `ComparingStringsWithoutStringComparison`: Preserves leading trivia when applying the code fix
+- `ThreadSleepInAsyncMethod`: Rewritten to use `IOperation`
+- `ThreadSleepInAsyncMethod`: Supports `async` lambdas
+
 ## [1.17.1] - 2023-01-02
 - `ElementaryMethodsOfTypeInCollectionNotOverridden`: Fixed an issue where a `NullReferenceException` would be thrown
 - `HttpClientInstantiatedDirectly`: Only triggers if a reference to `Microsoft.Extensions.Http` exists and `IHttpClientFactory` is within scope